requirements.txt
# Project Sentinel - AI-Powered UFDR Analysis Platform
# Core Dependencies for Case-Based Forensic Investigation Platform

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database (SQLAlchemy for case management)
sqlalchemy==2.0.23
alembic==1.13.1

# AI and ML Libraries - Core LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
langchain-openai==0.0.5
langchain-anthropic==0.0.4

# LangGraph for State-Based AI Workflows
langgraph==0.0.26
langgraph-checkpoint==1.0.0

# ML and Embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
scikit-learn==1.3.2

# Advanced LLM Integrations
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.7

# Future Multi-Modal Support (CLIP for images)
# clip-by-openai==1.0
# Pillow==10.1.0

# Vector Database and Embeddings
chromadb==0.4.18
faiss-cpu==1.7.4

# Graph Database (for advanced relationship analysis)
neo4j==5.15.0
py2neo==2021.2.4
networkx==3.2.1

# NLP and Entity Recognition
spacy==3.7.2
spacy-transformers==1.3.4
nltk==3.8.1
regex==2023.10.3

# Data Processing and Parsing
pandas==2.1.4
lxml==4.9.3
beautifulsoup4==4.12.2
xmltodict==0.13.0
openpyxl==3.1.2

# Frontend
streamlit==1.29.0
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and Monitoring
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security and Authentication
cryptography==41.0.8
bcrypt==4.1.2

# Async and Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Datetime and Timezone
python-dateutil==2.8.2
pytz==2023.3

# Image and Vision (for future multi-modal support)
Pillow==10.1.0
opencv-python==4.8.1.78

# HTTP Requests and APIs
httpx==0.25.2
requests==2.31.0

# Database Connectivity
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# JSON and Data Serialization
orjson==3.9.10